+++ src/Translator/src/Catalogue/CatalogueLoader.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Translator\Catalogue;
@@ -16,36 +9,26 @@ use Spiral\Translator\Catalogue;
 use Spiral\Translator\CatalogueInterface;
 use Spiral\Translator\Config\TranslatorConfig;
 use Symfony\Component\Finder\Finder;
-use Symfony\Component\Finder\SplFileInfo;
 
 final class CatalogueLoader implements LoaderInterface
 {
     use LoggerTrait;
 
-    /** @var TranslatorConfig */
-    private $config;
-
-    public function __construct(TranslatorConfig $config)
-    {
-        $this->config = $config;
+    public function __construct(
+        private readonly TranslatorConfig $config
+    ) {
     }
 
-    /**
-     * @inheritdoc
-     */
     public function hasLocale(string $locale): bool
     {
-        $locale = preg_replace('/[^a-zA-Z_]/', '', mb_strtolower($locale));
+        $locale = \preg_replace('/[^a-zA-Z_]/', '', \mb_strtolower($locale));
 
-        return is_dir($this->config->getLocaleDirectory($locale));
+        return \is_dir($this->config->getLocaleDirectory($locale));
     }
 
-    /**
-     * @inheritdoc
-     */
     public function getLocales(): array
     {
-        if (!is_dir($this->config->getLocalesDirectory())) {
+        if (!\is_dir($this->config->getLocalesDirectory())) {
             return [];
         }
 
@@ -61,12 +44,9 @@ final class CatalogueLoader implements LoaderInterface
         return $locales;
     }
 
-    /**
-     * @inheritdoc
-     */
     public function loadCatalogue(string $locale): CatalogueInterface
     {
-        $locale = preg_replace('/[^a-zA-Z_]/', '', mb_strtolower($locale));
+        $locale = \preg_replace('/[^a-zA-Z_]/', '', \mb_strtolower($locale));
         $catalogue = new Catalogue($locale);
 
         if (!$this->hasLocale($locale)) {
@@ -78,7 +58,7 @@ final class CatalogueLoader implements LoaderInterface
 
         foreach ($finder->getIterator() as $file) {
             $this->getLogger()->info(
-                sprintf(
+                \sprintf(
                     "found locale domain file '%s'",
                     $file->getFilename()
                 ),
@@ -86,11 +66,11 @@ final class CatalogueLoader implements LoaderInterface
             );
 
             //Per application agreement domain name must present in filename
-            $domain = strstr($file->getFilename(), '.', true);
+            $domain = \strstr($file->getFilename(), '.', true);
 
             if (!$this->config->hasLoader($file->getExtension())) {
                 $this->getLogger()->warning(
-                    sprintf(
+                    \sprintf(
                         "unable to load domain file '%s', no loader found",
                         $file->getFilename()
                     ),