+++ src/Translator/src/Exception/LocaleException.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Translator\Exception;
@@ -16,20 +9,12 @@ namespace Spiral\Translator\Exception;
  */
 class LocaleException extends TranslatorException
 {
-    /**
-     * @var string
-     */
-    protected $locale;
-
-    /**
-     * @param string     $locale
-     * @param int        $code
-     * @param \Exception $previous
-     */
-    public function __construct($locale, $code = 0, $previous = null)
-    {
-        $this->locale = $locale;
-        parent::__construct("Undefined locale '{$locale}'", $code, $previous);
+    public function __construct(
+        protected string $locale,
+        int $code = 0,
+        \Throwable $previous = null
+    ) {
+        parent::__construct(\sprintf('Undefined locale \'%s\'', $locale), $code, $previous);
     }
 
     public function getLocale(): string