+++ src/Translator/src/Translator.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Translator;
@@ -23,42 +16,24 @@ use Symfony\Component\Translation\IdentityTranslator;
  */
 final class Translator implements TranslatorInterface, SingletonInterface
 {
-    /** @var TranslatorConfig */
-    private $config;
-
-    /** @var string */
-    private $locale = '';
-
-    /** @var IdentityTranslator @internal */
-    private $identityTranslator;
-
-    /** @var CatalogueManagerInterface */
-    private $catalogueManager;
+    private string $locale;
 
     public function __construct(
-        TranslatorConfig $config,
-        CatalogueManagerInterface $catalogueManager,
-        IdentityTranslator $identityTranslator = null
+        private readonly TranslatorConfig $config,
+        private readonly CatalogueManagerInterface $catalogueManager,
+        /** @internal */
+        private readonly IdentityTranslator $identityTranslator = new IdentityTranslator()
     ) {
-        $this->config = $config;
-        $this->identityTranslator = $identityTranslator ?? new IdentityTranslator();
-        $this->catalogueManager = $catalogueManager;
-
         $this->locale = $this->config->getDefaultLocale();
         $this->catalogueManager->load($this->locale);
     }
 
-    /**
-     * @inheritdoc
-     */
     public function getDomain(string $bundle): string
     {
         return $this->config->resolveDomain($bundle);
     }
 
     /**
-     * @inheritdoc
-     *
      * @throws LocaleException
      */
     public function setLocale(string $locale): void
@@ -71,30 +46,22 @@ final class Translator implements TranslatorInterface, SingletonInterface
         $this->catalogueManager->load($locale);
     }
 
-    /**
-     * @inheritdoc
-     */
     public function getLocale(): string
     {
         return $this->locale;
     }
 
-    /**
-     * @inheritdoc
-     */
     public function getCatalogueManager(): CatalogueManagerInterface
     {
         return $this->catalogueManager;
     }
 
     /**
-     * {@inheritdoc}
-     *
      * Parameters will be embedded into string using { and } braces.
      *
      * @throws LocaleException
      */
-    public function trans(string $id, array $parameters = [], $domain = null, $locale = null): string
+    public function trans(string $id, array $parameters = [], string $domain = null, string $locale = null): string
     {
         $domain = $domain ?? $this->config->getDefaultDomain();
         $locale = $locale ?? $this->locale;
@@ -105,8 +72,6 @@ final class Translator implements TranslatorInterface, SingletonInterface
     }
 
     /**
-     * {@inheritdoc}
-     *
      * Default symfony pluralizer to be used. Parameters will be embedded into string using { and }
      * braces. In addition you can use forced parameter {n} which contain formatted number value.
      *
@@ -114,12 +79,12 @@ final class Translator implements TranslatorInterface, SingletonInterface
      * @throws PluralizationException
      */
     public function transChoice(
-        $id,
-        $number,
+        string $id,
+        string|int $number,
         array $parameters = [],
-        $domain = null,
-        $locale = null
-    ) {
+        string $domain = null,
+        string $locale = null
+    ): string {
         $domain = $domain ?? $this->config->getDefaultDomain();
         $locale = $locale ?? $this->locale;
 
@@ -150,7 +115,6 @@ final class Translator implements TranslatorInterface, SingletonInterface
      * Input: Hello {name}! Good {time}! + ['name' => 'Member', 'time' => 'day']
      * Output: Hello Member! Good Day!
      *
-     * @param string $string
      * @param array  $values Arguments (key => value). Will skip unknown names.
      * @param string $prefix Placeholder prefix, "{" by default.
      * @param string $postfix Placeholder postfix, "}" by default.
@@ -163,10 +127,10 @@ final class Translator implements TranslatorInterface, SingletonInterface
     ): string {
         $replaces = [];
         foreach ($values as $key => $value) {
-            $value = (is_array($value) || $value instanceof \Closure) ? '' : $value;
+            $value = (\is_array($value) || $value instanceof \Closure) ? '' : $value;
 
-            if (is_object($value)) {
-                if (method_exists($value, '__toString')) {
+            if (\is_object($value)) {
+                if (\method_exists($value, '__toString')) {
                     $value = $value->__toString();
                 } else {
                     $value = '';
@@ -176,7 +140,7 @@ final class Translator implements TranslatorInterface, SingletonInterface
             $replaces[$prefix . $key . $postfix] = $value;
         }
 
-        return strtr($string, $replaces);
+        return \strtr($string, $replaces);
     }
 
     /**
@@ -184,8 +148,8 @@ final class Translator implements TranslatorInterface, SingletonInterface
      */
     public static function isMessage(string $string): bool
     {
-        return substr($string, 0, 2) == self::I18N_PREFIX
-            && substr($string, -2) == self::I18N_POSTFIX;
+        return \substr($string, 0, 2) == self::I18N_PREFIX
+            && \substr($string, -2) == self::I18N_POSTFIX;
     }
 
     /**